requirements.txt
cachetools==5.3.2
google-auth==2.24.0
google-auth-oauthlib==1.1.0
gspread==5.12.1
oauthlib==3.2.2
pyasn1==0.5.1
pyasn1-modules==0.3.0
requests-oauthlib==1.3.1
rsa==4.9
